backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
pydantic
google-auth
requests
python-jose[cryptography]
passlib[bcrypt]